+++ src/main/java/com/blog/web/form/CategoryForm.java
@@ -11,5 +11,6 @@ public class CategoryForm {
     @NotBlank(message = "카테고리 명은 필수입니다.")
     private String title;
     @NotBlank(message = "목록 순서는 필수입니다.")
-    private Long listOrder;
+    private Integer listOrder;
+
 }
